UIAutomationLib/requirements.txt
robotframework==7.0
robotframework-seleniumlibrary==6.3.0
selenium==4.21.0